backend/requirements.txt
fastapi==0.111.0
python-multipart==0.0.18
uvicorn[standard]==0.30.6
Pillow
google-generativeai==0.7.2
pydantic==2.9.2
